python-server/requirements.txt
requests
python-dotenv
gradio
pypdf
openai
openai-agents
pyinstaller
uvicorn
fastapi
crewai[tools]
crewai-tools
